feat(pkg/executor): move job-executor logic to pkg
This makes it easier for library users to enhance their own
applications to behave as a job-executor instead of running
./cmd/job-executor.

The readme has been updated on these steps.
